src/app/app.component.spec.ts
import { TestBed, async ,ComponentFixture} from '@angular/core/testing';
import { By }              from '@angular/platform-browser';
import { DebugElement }    from '@angular/core';
import { AppComponent } from './app.component';
import {FormsModule} from '@angular/forms';
import { QuotaComponent } from './quota/quota.component';
describe('AppComponent', () => {
  let component: AppComponent;
  let fixture: ComponentFixture<AppComponent>;
  let de:      DebugElement;
  let el:      HTMLElement;
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [FormsModule],
      declarations: [AppComponent,QuotaComponent],
    }).compileComponents();
  }));
  beforeEach(() => {
    fixture = TestBed.createComponent(AppComponent);
    component = fixture.componentInstance;
    de = fixture.debugElement.query(By.css('.container'));
    el = de.nativeElement;
  });


  it('should create the app', async(() => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  }));
  it(`should have as title 'app'`, async(() => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('app');
  }));
  it('should render title in a h1 tag', async(() => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('Welcome to app!');
  }));

  it('should show the BannerTitle', () => {
    fixture.detectChanges();
    const content = el.textContent;
    expect(content).toContain('CREW LEAVE');
  });

  it('should show the Show Quota Button', () => {
    fixture.detectChanges();
    de = fixture.debugElement.query(By.css('.btn-toolbar pull-right'));
    const content = el.textContent;
    expect(content).toContain('Show Quota');
  });
});
